partial UI integration

+++ src/main/java/edu/northeastern/cs5200/person/GymLeader/GymLeaderRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GymLeaderRepository extends CrudRepository<GymLeader, Integer> {
+public interface GymLeaderRepository extends CrudRepository<GymLeader, Long> {
 }
