+++ src/main/java/edu/northeastern/cs5200/person/trainer/TrainerRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TrainerRepository extends CrudRepository<Trainer, Integer> {
+public interface TrainerRepository extends CrudRepository<Trainer, Long> {
 
 }
